Duplicate checking based on CRC
- Instead of looking at a single byte of data, check the CRC on potential duplicate packets

+++ src/nrf_to_nrf.h
@@ -378,7 +378,7 @@ private:
   bool ackPayloadAvailable;
   uint8_t ackAvailablePipeNo;
   uint8_t lastPacketCounter;
-  uint8_t lastData;
+  uint16_t lastData;
   bool dynamicAckEnabled;
   uint8_t ARC;
   uint8_t addressWidth;
